fix: medusa-plugin-whatsapp readme
